+++ include/color.h
@@ -6,14 +6,14 @@
 // Struct for an RGB color representation //
 typedef struct s_color
 {
-	double	r; // Value from 0 to 1
-	double	g; // Value from 0 to 1
-	double	b; // Value from 0 to 1
+	float	r; // Value from 0 to 1
+	float	g; // Value from 0 to 1
+	float	b; // Value from 0 to 1
 }	t_color;
 
 // Functions //
 
-t_color		make_color(double r, double g, double b);
+t_color		make_color(float r, float g, float b);
 uint32_t	color_to_int(t_color *color);
 
 #endif // COLOR_H
