+++ include/light.h
@@ -7,7 +7,7 @@
 typedef struct s_light
 {
 	t_point3		position;
-	double			ratio;	// Range [0, 1.0]
+	float			ratio;	// Range [0, 1.0]
 	t_color			color;	// RGB color
 	struct s_light	*next;
 }	t_light;
