+++ src/scene/camera.c
@@ -5,7 +5,7 @@
 
 #include <stdio.h>
 
-void	create_camera(t_point3 position, t_vec3 direction, double fov)
+void	create_camera(t_point3 position, t_vec3 direction, float fov)
 {
 	t_camera	*camera;
 
@@ -37,8 +37,8 @@ t_vec3	camera_to_viewport(int x, int y)
 	return (
 		(t_vec3)
 		{
-			.x = x * (1.0 / (double)WIN_WIDTH),
-			.y = y * (1.0 / (double)WIN_HEIGHT),
+			.x = x * (1.0 / (float)WIN_WIDTH),
+			.y = y * (1.0 / (float)WIN_HEIGHT),
 			.z = 1.0
 		}
 	);
